+++ Programming/lab5/src/main/java/org/example/commands/UpdateById.java
@@ -6,10 +6,11 @@ import org.example.collections.forms.StandardForm;
 import org.example.collections.forms.nonUserMode.LabWorkFormNonUserMode;
 import org.example.managers.CollectionManager;
 import org.example.managers.CommandManager;
+import org.example.managers.IdManager;
 import org.example.utility.Console;
 
 /**
- * Команда update_id
+ * Команда, которая обновляет элемент по id
  */
 public class UpdateById implements Command {
     private final CollectionManager collectionManager;
@@ -24,24 +25,39 @@ public class UpdateById implements Command {
 
     /**
      * Обновляет элемент коллекции по id
+     *
      * @param arguments Аргументы
      */
     @Override
     public void execute(String[] arguments) {
         StandardForm<LabWork> form;
-        long id = Long.parseLong(arguments[0]);
-        collectionManager.deleteById(id);
-        if (commandManager.isUserMode()) {
-            form = new LabWorkForm(console);
+        if (collectionManager.isEmpty()) {
+            console.println("Коллекция пуста");
         } else {
-            LabWorkFormNonUserMode nonUserMode = new LabWorkFormNonUserMode(console);
-            nonUserMode.setArgs(arguments);
-            form = nonUserMode;
+            long id = Long.parseLong(arguments[0]);
+            if (IdManager.containsId(id)) {
+                collectionManager.deleteById(id);
+                if (commandManager.isUserMode()) {
+                    form = new LabWorkForm(console);
+                } else {
+                    LabWorkFormNonUserMode nonUserMode = new LabWorkFormNonUserMode(console);
+                    nonUserMode.setArgs(arguments);
+                    form = nonUserMode;
+                }
+                LabWork labWork = form.build();
+                labWork.setId(id);
+                if (labWork != null) {
+                    collectionManager.addToCollection(labWork);
+                    console.println("Объект обновлен!");
+                } else {
+                    console.println("Объект не обновлен");
+                }
+
+            } else {
+                console.println("Такого id не существует");
+            }
         }
-        LabWork labWork = form.build();
-        labWork.setId(id);
-        collectionManager.addToCollection(labWork);
-        console.println("Объект обновлен!");
+
     }
 
     @Override
